Fix clang-format

+++ recipe/install_clang_tools.sh
@@ -6,6 +6,9 @@ cd $PREFIX
 rm -rf lib/cmake include lib/lib*.a
 MAJOR_VERSION=$(echo ${PKG_VERSION} | cut -f1 -d".")
 for f in ${PREFIX}/bin/clang-*; do
+    if [[ "$(basename $f)" == clang-format-* ]]; then
+        continue
+    fi
     rm -f ${PREFIX}/bin/$(basename $f)-${MAJOR_VERSION}
     mv $f ${PREFIX}/bin/$(basename $f)-${MAJOR_VERSION};
     ln -s ${PREFIX}/bin/$(basename $f)-${MAJOR_VERSION} $f;
